docs(frontend): getting params from TFHE-rs
